Introduce the concept of apps
This is a rather big change that should allow you to easily customize
how satchless behaves. We are starting with product and category views
and will continue to introduce the apps to the rest of the framework.

For now the biggest change is how you include the URLs in your urls.py
file:

    from satchless.category.app import product_app
    #...
    urlpatterns = patterns('',
        url(r'^product/', include(product_app.urls))
    )